Move `CSnapIDPool` to separate compilation unit
